+++ src/views/admin-configuration/gold-config.vue
@@ -29,7 +29,11 @@
             <el-table-column label="使用状态">
               <template slot-scope="scope">{{ scope.row.state | changeStateName }}</template>
             </el-table-column>
-            <el-table-column prop="createTime" label="创建时间" />
+            <el-table-column label="创建时间">
+              <template slot-scope="scope">
+                {{ scope.row.createTime | toTimeSpan }}
+              </template>
+            </el-table-column>
             <el-table-column prop="createStaff" label="创建人" />
             <el-table-column label="操作" width="190">
               <template slot-scope="scope">
@@ -112,33 +116,47 @@ export default {
      * @description 删除金币配置
      */
     async deleteData (cardItem) {
-      try {
-        this.$store.dispatch('setting/showLoading', this.routeName)
-        const reqData = { staffCardId: cardItem.id }
-        await OperationManage.deleteCard(reqData)
-        this.$newMessage.success('删除成功')
-        this.getGoldConfigList()
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-      } catch (error) {
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-        console.error(error)
-      }
+      this.$confirm('确认删除该金币奖励吗？', '', {
+        confirmButtonText: '确认',
+        cancelButtonText: '取消',
+        type: 'warning',
+        center: true
+      }).then(async () => {
+        try {
+          this.$store.dispatch('setting/showLoading', this.routeName)
+          const reqData = { staffCardId: cardItem.id }
+          await OperationManage.deleteCard(reqData)
+          this.$newMessage.success('删除成功')
+          this.getGoldConfigList()
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+        } catch (error) {
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+          console.error(error)
+        }
+      }).catch(() => {})
     },
     /**
      * @description 提前结束
      */
     async closeCard (cardItem) {
-      try {
-        this.$store.dispatch('setting/showLoading', this.routeName)
-        const reqData = { staffCardId: cardItem.id }
-        await OperationManage.closeCard(reqData)
-        this.$newMessage.success('提前结束')
-        this.getGoldConfigList()
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-      } catch (error) {
-        this.$store.dispatch('setting/hiddenLoading', this.routeName)
-        console.error(error)
-      }
+      this.$confirm('确认提前结束吗？', '', {
+        confirmButtonText: '确认',
+        cancelButtonText: '取消',
+        type: 'warning',
+        center: true
+      }).then(async () => {
+        try {
+          this.$store.dispatch('setting/showLoading', this.routeName)
+          const reqData = { staffCardId: cardItem.id }
+          await OperationManage.closeCard(reqData)
+          this.$newMessage.success('提前结束')
+          this.getGoldConfigList()
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+        } catch (error) {
+          this.$store.dispatch('setting/hiddenLoading', this.routeName)
+          console.error(error)
+        }
+      }).catch(() => {})
     },
     /**
      * @description 获取金币配置列表
